[dev] Add the support new compression type for the decompressor
RLE bits, sub pixel bits and image size are not constant anymore.
So a new structure EerFrameSettings has been create to propagate all this
information trought de depressor algorithm.
The EerFrameSettings are created with the first EER frame or define with
the previous constant value in case of ECC mode.

Static variable uses by the defect masking class has been move in this one.
The change are not apply on the defect masking because it done in the
acquistion pipeline for 2K images.